python/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic>=2.7,<3
python-dotenv>=1.0,<2
httpx>=0.27,<0.28

# testing
pytest>=8.0,<9
anyio>=3.7,<5